+++ tests/test_toy_datasets.py
@@ -1,8 +1,8 @@
-from sopa.utils.data import blobs, uniform
+from sopa.utils.data import blobs, toy_dataset
 
 
-def test_make_uniform():
-    assert uniform(length=512) is not None
+def test_make_toy_dataset():
+    assert toy_dataset(length=512) is not None
 
 
 def test_make_blobs():